pynipap: Merged to single file
To avoid hogging the global name "tracing", merged tracing.py into
pynipap.py.
